refactor: move token utils to internal utils package
This commit moves the token utility functions from the `internal/tokens` package to the `internal/utils` package. This change improves the organization of the codebase and centralizes general utility functions in a common location.

The following files have been moved:

- `internal/tokens/token.go` -> `internal/utils/token.go`
- `auth/tests/token_generation_test.go` -> `internal/utils/tests/token_generation_test.go`

The `auth` package and tests have been updated to import these functions from the new location.