feat: add new:feat command to generate feature files
This commit introduces a new command `new:feat` that allows developers to quickly generate feature files within specified modules (api or database).

The command takes the feature name as an argument and creates a directory with the feature name inside the target module. It also generates a basic Go file with the feature name as the package name.

The target module can be specified using the `--api` or `--database` flags. If no flag is provided, the default target is the `api` module.